slight rework of testing filenames
